% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE.R
\name{SFEcmprsk}
\alias{SFEcmprsk}
\title{SurvELM ELMCox}
\usage{
SFEcmprsk(x, y, lambda, beta = dim(x)[2])
}
\arguments{
\item{x}{The covariates(predictor variables) of training data.}

\item{y}{Survival time and censored status of training data. First column be "time" and the second column be "status"}

\item{lambda}{Penalty Parameters.}

\item{beta}{Initial values for beta. set all to 1 as default for the Lasso PSH model.}
}
\value{
A list of elements
  \tabular{ll}{
      \code{delindex}    \tab  A vector of index in x that to be discarded \cr
      \code{numdel} \tab  The number of discarded features. \cr
  }
}
\description{
SAFE Feature Screening for Lasso and Adaptive Lasso Models
}
\examples{
set.seed(123)
require(survival)
}
\author{
Hong Wang
}
